try and get travis ci working with ssh
